raindrops: standard test cases, test version, stub
