+++ raindrops/cases_test.go
@@ -1,0 +1,25 @@
+package raindrops
+
+// Source: exercism/x-common
+// Commit: 3b07e53 Merge pull request #117 from mikeyjcat/add-raindrops-json
+
+var tests = []struct {
+	input    int
+	expected string
+}{
+	{1, "1"},
+	{3, "Pling"},
+	{5, "Plang"},
+	{7, "Plong"},
+	{6, "Pling"},
+	{9, "Pling"},
+	{10, "Plang"},
+	{14, "Plong"},
+	{15, "PlingPlang"},
+	{21, "PlingPlong"},
+	{25, "Plang"},
+	{35, "PlangPlong"},
+	{49, "Plong"},
+	{52, "52"},
+	{105, "PlingPlangPlong"},
+}
