+++ raindrops/raindrops.go
@@ -1,0 +1,9 @@
+// +build !example
+
+package raindrops
+
+const TestVersion = 1
+
+func Convert(int) string
+
+// The test program has a benchmark too.  How fast does your Convert convert?
